+++ src/rex.deploy/src/rex/deploy/link.py
@@ -8,7 +8,8 @@ from .fact import Fact, LabelVal, QLabelVal, TitleVal, label_to_title
 from .image import SET_DEFAULT
 from .meta import ColumnMeta, TableMeta
 from .sql import (mangle, sql_add_column, sql_drop_column,
-        sql_comment_on_column, sql_add_foreign_key_constraint)
+        sql_comment_on_column, sql_add_foreign_key_constraint,
+        sql_create_index)
 
 
 class LinkFact(Fact):
@@ -153,7 +154,8 @@ class LinkFact(Fact):
         if column.is_not_null != self.is_required:
             raise Error("Detected column with mismatched"
                         " NOT NULL constraint:", column)
-        # Create a `FOREIGN KEY` constraint if necessary.
+        # Create a `FOREIGN KEY` constraint and an associated index
+        # if necessary.
         if not any(foreign_key
                    for foreign_key in table.foreign_keys
                    if list(foreign_key) == [(column, target_column)]):
@@ -166,6 +168,9 @@ class LinkFact(Fact):
             table.add_foreign_key(self.constraint_name, [column],
                                   target_table, [target_column],
                                   on_delete=SET_DEFAULT)
+            driver.submit(sql_create_index(self.constraint_name,
+                    self.table_name, [self.name]))
+            # FIXME: keep track of indexes?
         # Store the original link label and the link title.
         meta = ColumnMeta.parse(column)
         preferred_label = self.name[:-2].rstrip(u'_') \